feat: implementation of evaluation registration logic
- addition of validators to cover each business rule
- addition of unit tests to cover each business rule

+++ src/main/java/br/com/meli/dhprojetointegrador/entity/PurchaseOrderEvaluation.java
@@ -19,7 +19,7 @@ public class PurchaseOrderEvaluation implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String comment;
-    private Short rating;
+    private Integer rating;
 
     @ManyToOne
     @JoinColumn(name = "fk_purchase_order_id")
